# docs/technical/openapi.yaml
# Generated by `pnpm docs` from API_ROUTES in @bs/protocol. Do not edit.
# Schemas are named after the TypeScript types in packages/protocol/src/wire.ts,
# which the technical specification embeds in full.
openapi: 3.1.0
info:
  title: Orbital Battleships Command API
  version: 1.0.0
  description: >-
    Server-authoritative Battleship. The alien deployment is never returned
    while a campaign is in progress.
servers:
  - url: /
security:
  - playerToken: []
paths:
  /api/matches:
    post:
      summary: "Start a campaign. Deploys the player fleet, or accepts one, and the invader fleet."
      operationId: postApiMatches
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMatchRequest'
      responses:
        '200':
          description: "CreateMatchResponse"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateMatchResponse'
        '400':
          description: 'Malformed request or illegal move'
  /api/matches/{id}:
    get:
      summary: "Current state of a campaign, from the player’s point of view."
      operationId: getApiMatchesId
      security:
        - playerToken: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: "MatchView"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MatchView'
        '400':
          description: 'Malformed request or illegal move'
        '401':
          description: 'Missing player token'
        '403':
          description: 'Not your campaign'
        '404':
          description: 'No such campaign'
  /api/matches/{id}/fire:
    post:
      summary: "Fire at one cell of the invader grid; the invader replies in the same call."
      operationId: postApiMatchesIdFire
      security:
        - playerToken: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FireRequest'
      responses:
        '200':
          description: "MatchView"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MatchView'
        '400':
          description: 'Malformed request or illegal move'
        '401':
          description: 'Missing player token'
        '403':
          description: 'Not your campaign'
        '404':
          description: 'No such campaign'
  /api/matches/{id}/resign:
    post:
      summary: "Concede the campaign. The invader fleet is revealed."
      operationId: postApiMatchesIdResign
      security:
        - playerToken: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: "MatchView"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MatchView'
        '400':
          description: 'Malformed request or illegal move'
        '401':
          description: 'Missing player token'
        '403':
          description: 'Not your campaign'
        '404':
          description: 'No such campaign'
  /api/me/progress:
    get:
      summary: "Career record and rank for the bearer token."
      operationId: getApiMeProgress
      security:
        - playerToken: []
      responses:
        '200':
          description: "ProgressResponse"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProgressResponse'
        '400':
          description: 'Malformed request or illegal move'
        '401':
          description: 'Missing player token'
  /api/leaderboard:
    get:
      summary: "The shared board of best campaigns, seeded captains included."
      operationId: getApiLeaderboard
      security: []
      responses:
        '200':
          description: "LeaderboardResponse"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LeaderboardResponse'
        '400':
          description: 'Malformed request or illegal move'
  /api/fleet:
    get:
      summary: "Fleet roster, grid size, scoring table and difficulty doctrines. Public reference data."
      operationId: getApiFleet
      security: []
      responses:
        '200':
          description: "ReferenceResponse"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReferenceResponse'
        '400':
          description: 'Malformed request or illegal move'
  /api/health:
    get:
      summary: "Liveness probe with the deployed version."
      operationId: getApiHealth
      security: []
      responses:
        '200':
          description: "HealthResponse"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'
        '400':
          description: 'Malformed request or illegal move'
components:
  securitySchemes:
    playerToken:
      type: apiKey
      in: header
      name: x-player-token
      description: 32-byte hex token issued by POST /api/matches.
  schemas:
    Placement:
      type: object
      required: [hull, origin, orientation]
      properties:
        hull:
          type: string
          enum: [battleship-1, cruiser-1, cruiser-2, destroyer-1, destroyer-2, destroyer-3, submarine-1, submarine-2, submarine-3, submarine-4]
        origin:
          type: integer
          minimum: 0
          maximum: 99
        orientation:
          type: string
          enum: [horizontal, vertical]
    Shot:
      type: object
      required: [cell, outcome]
      properties:
        cell:
          type: integer
          minimum: 0
          maximum: 99
        outcome:
          type: string
          enum: [miss, hit, sunk]
        hull:
          type: string
          description: Present only when the outcome is sunk.
    CreateMatchRequest:
      type: object
      required: [difficulty]
      properties:
        difficulty:
          type: string
          enum: [scout, raider, overmind]
        fleet:
          type: array
          description: Optional manual deployment; generated server-side when absent.
          items:
            $ref: '#/components/schemas/Placement'
    FireRequest:
      type: object
      required: [cell]
      properties:
        cell:
          type: integer
          minimum: 0
          maximum: 99
    MatchView:
      type: object
      description: >-
        The campaign as the player may see it. alienFleet is present only
        once the campaign has finished.
      required: [matchId, difficulty, status, turn, defence, offence, stats, score, log]
      additionalProperties: true
    CreateMatchResponse:
      type: object
      required: [match, playerToken]
      properties:
        match:
          $ref: '#/components/schemas/MatchView'
        playerToken:
          type: string
    ProgressResponse:
      type: object
      required: [progress, rank, recent]
      additionalProperties: true
    ReferenceResponse:
      type: object
      description: Fleet, scoring and doctrine reference data.
      additionalProperties: true
    HealthResponse:
      type: object
      required: [status, version]
      properties:
        status:
          type: string
        version:
          type: string
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string